+++ app/scripts/contentScript.js
@@ -9,6 +9,7 @@ import * as gl from './lib/generic-lib';
 import * as gdl from './lib/generic-dom-lib';
 import * as notifyButton from './lib/notifyButton';
 import * as gcl from './lib/generic-comment-lib';
+import * as TT from './lib/timetable';
 import './updatenotify.js';
 
 // edge対応
@@ -70,8 +71,6 @@ var isFuga = false; //利用したくない機能
 var cmblockia = 1; //コメント欄が無効になってからCM処理までのウェイト(+1以上)
 var cmblockib = -1; //有効になってから解除までのウェイト(-1以下)
 var panelopenset = [[1, 1, 1], [0, 0, 0], [0, 0, 0], [0, 0, 0]]; //head,foot,sideの開閉設定[全閉,info開,chli開,come開] 0:非表示 1:5秒で隠す 2:常に表示
-//settings.allowChannelNames = ["abema-news", "abema-special", "special-plus", "special-plus-2", "special-plus-3", "special-plus-4", "special-plus-5", "special-plus-6", "drama", "asia-drama", "reality-show", "mtv-hits", "space-shower", "documentary", "variety", "pet", "club-channel", "commercial", "anime24", "midnight-anime", "oldtime-anime", "family-anime", "new-anime", "yokonori-sports", "hiphop", "soccer", "fighting-sports", "fighting-sports2", "golf", "fishing", "shogi", "mahjong"];
-settings.allowChannelNames = [];
 
 var disableExtVersion = ''; //拡張機能の動作を停止するバージョン
 
@@ -124,13 +123,8 @@ console.log('script loaded');
         value.highlightComePower !== undefined
             ? Number(value.highlightComePower)
             : settings.highlightComePower;
-    settings.allowChannelNames =
-        value.allowChannelNames !== undefined
-            ? value.allowChannelNames.split(',')
-            : settings.allowChannelNames;
     //無視する設定
     settings.isInpWinBottom = false;
-    //settings.isComelistNG = false;
 })();
 
 var currentLocation = window.location.href;
@@ -183,12 +177,6 @@ let EXvideoarea;
 var EXvolume;
 var EXfullscr;
 var EXcomemodule; //Twitterボタンや投稿ボタンを含む要素 入力欄をフォーカスしたときにセットし放送画面以外ではnull
-var EXTThead; //timetableのヘッダ部分(チャンネル,日付の親)
-var EXTTbody; //timetableのボディ(チャンネル,日付の親)
-var EXTTbodyS; //bodyのスクロール担当(sideL,Rの兄弟)
-var EXTTsideR; //timetableの右の番組詳細
-var EXTTsideL; //timetableの左の番組一覧
-var EXTTtime; //timetableの時間列(縦長の緑のやつ)
 var cmblockcd = 0; //カウント用
 var comeRefreshing = false; //コメ欄自動開閉中はソートを実行したいのでコメント更新しない用
 var comeFastOpen = false;
@@ -220,7 +208,6 @@ var isComelistMouseDown = false;
 var movieWidth = 0; //.TVContainer__resize-screenの大きさ(onresize発火用)
 //var movieHeight = 0;
 var waitingforResize = false; //waitforresizeの複数起動防止用
-var allowChannelNum = []; //Namesを元にした表示列番号
 //var movieWidth2 = 0; //video.parentの大きさ(onresize発火用)
 var oldWindowState = 'normal'; // フルスクリーン切り替え前のウィンドウのstate
 var isTootEnabled = false; //コメントのトゥート有効か
@@ -235,18 +222,9 @@ var commentObserver = new MutationObserver(function(mutations) {
     ); //injection.jsでdata属性のセットが終わるまで待つ
 }); //コメント欄DOM監視
 var isFirstComeAnimated = false; //最初に既存のコメがanimationしたか 最初に既存のコメが一気に画面に流れてくるのを防ぐためのフラグ
-var timetableRunning = false; //番組表表示時の10分インターバル
 var audibleReloadCount = -1;
 var isSoundFlag = true; //音が出ているか soundSet(isSound)のisSoundを保持したり音量クリック時にミュートチェック
-var timetableGrabbing = {
-    value: false,
-    cx: 0,
-    cy: 0,
-    test: false,
-    sx: 0,
-    sy: 0,
-    scrolled: false
-}; //番組表を掴む
+
 var comelistClasses = {
     comment: '',
     stabled: '',
@@ -257,7 +235,6 @@ var comelistClasses = {
     posttime: '',
     newhilight: ''
 };
-var timetableClasses = { arrow: '', timebar: '' }; //ページ遷移直後に取得できないので初回取得時に保持する getSingleSelectorの結果を入れるので使用時は.を付けない
 var currentVersion = chrome.runtime.getManifest().version;
 var resizeEventTimer = 0; //ウィンドウリサイズイベント用のタイマー
 var isBottomScrolled = false; //コメ欄逆順時初回で下にスクロールしたか
@@ -302,928 +279,6 @@ function onairCleaner() {
     //DOM監視停止
     commentObserver.disconnect();
 }
-function allowChannelNumMaker() {
-    //console.log("allowChannelNumMaker");
-    if (settings.allowChannelNames.length == 0) return 2;
-    var u = 'https://abema.tv/timetable/channels/';
-    var eaHead = $(EXTThead).children('a');
-    if (eaHead.length == 0) return -2;
-    var n = [];
-    for (var i = 0, f, h, c; i < eaHead.length; i++) {
-        if ((h = eaHead.eq(i).prop('href')) && h.indexOf(u) == 0) {
-            c = h.replace(u, '');
-            //console.log("c="+c);
-            if (settings.allowChannelNames.includes(c)) {
-                n.push(i);
-            }
-        }
-    }
-    if (n.length == 0) return -1;
-    var b = false;
-    if (n.length == allowChannelNum.length) {
-        b = true;
-        for (let i = 0; i < n.length; i++) {
-            if (n[i] != allowChannelNum[i]) {
-                b = false;
-                break;
-            }
-        }
-    }
-    allowChannelNum = n;
-    if (b) {
-        return 0;
-    } else {
-        return 1;
-    }
-}
-function timetableCss() {
-    //console.log("timetableCss");
-    $('head>link[title="usermade"]').remove();
-    var t = '';
-    var ts = '';
-    var to, tp;
-    var selBody, selHead, selTime, selPTitle, selBodyS;
-    var eo, ep;
-    var m;
-    var alt = false;
-
-    //番組タイトルが少ない状態だとうまく取れないが今の遅延適用される状態(ch列が少し溜まるまでこのcssが生成されない)のがうまいこと効いてる
-    selPTitle = getTTProgramTitleClass();
-    if (!selPTitle) {
-        console.log('?番組タイトルclass ' + to);
-        selPTitle = alt ? '.ok_bq' : '';
-    } else selPTitle = '.' + selPTitle;
-
-    selBody = dl.getElementSingleSelector(EXTTbody);
-    if ($(selBody).length != 1) {
-        console.log('?EXTTbody ' + selBody);
-        selBody = alt ? '.pi_pk' : '';
-    }
-
-    if (settings.isExpandFewChannels) {
-        //横長さ制限があるのはfuturetitleのpだけだが仕様変更に備えて全てのpに適用しておく
-        //futuretitleだけに適用する場合はgetTTTimeClassとかでがんばる
-        if (selBody) t += selBody + ' p{width:100%;padding-right:1em;}';
-
-        //横幅100%から左端～時間軸の右端と右のスクロールバーの分(適当)を引いた幅にする
-        //真面目にやるならexttbodyから上がっていってoverflowがvisibleでない要素(またはsideLの兄弟)のclientWidthを取ればスクロールバーまでバッチリ取れるかも
-        m = EXTTtime ? EXTTtime.getBoundingClientRect().right + 19 : 265;
-        ts =
-            'width:calc((100vw - ' +
-            m +
-            'px) / ' +
-            allowChannelNum.length +
-            ')!important;min-width:176px;';
-    }
-    if (settings.isChTimetableBreak && selPTitle) {
-        t += selPTitle + '{word-break:break-word;}';
-    }
-
-    var c = getInfo.determineUrl();
-    if (c == getInfo.URL_DATETABLE) {
-        if (allowChannelNum.length > 0) {
-            selHead = dl.getElementSingleSelector(EXTThead);
-            if ($(selHead).length != 1) {
-                console.log('?EXTThead ' + selHead);
-                selHead = alt ? '.qR_qT ' : '';
-            }
-            if (selHead) t += selHead + '>a{display:none;}';
-            if (selBody) t += selBody + '>div{display:none;}';
-            for (var i = 0, j; i < allowChannelNum.length; i++) {
-                j = allowChannelNum[i] + 1;
-                if (selHead)
-                    t +=
-                        selHead +
-                        '>a:nth-child(' +
-                        j +
-                        '){display:unset!important;' +
-                        ts +
-                        '}';
-                if (selBody)
-                    t +=
-                        selBody +
-                        '>div:nth-child(' +
-                        j +
-                        '){display:unset!important;' +
-                        ts +
-                        '}';
-            }
-        }
-        if (!timetableClasses.timebar) {
-            selTime = getTTtimebarElement(true);
-            if ($(selTime).length != 1) {
-                console.log('?date-bar ' + selTime);
-                selTime = alt ? '.i__j3' : '';
-            }
-        } else selTime = timetableClasses.timebar;
-        if (selTime) {
-            timetableClasses.timebar = selTime;
-            t += selTime + '{pointer-events:none;}';
-        }
-    }
-
-    if (settings.isHideArrowButton) {
-        if (!timetableClasses.arrow) {
-            to = getTTLRArrowContainerElement(true);
-            if ($(to).length != 1) {
-                console.log('?Arrowbutton ' + to);
-                to = alt ? '.i__jw' : '';
-            }
-        } else to = timetableClasses.arrow;
-        if (to) {
-            timetableClasses.arrow = to;
-            t += to + '{visibility:hidden;opacity:0;pointer-events:none;}';
-        }
-    }
-    if (settings.isExpandLastItem) {
-        if (selBody)
-            t +=
-                selBody +
-                '>div{height:unset;min-height:' +
-                ($(EXTTtime).height() || 4320) +
-                'px;}'; //各列の縦長さ制限を外す
-        if (selBody)
-            t += selBody + '>div>*:last-child>article{min-height:43px;}';
-    }
-    if (settings.isHideTodayHighlight) {
-        t += '[class*="styles__popup-container___"]{display:none;}'; //todo
-    }
-    if (settings.isChTimetablePlaybutton) {
-        t += '.playbutton:hover{background-color:yellow;}';
-    }
-    if (settings.isChTimetableWeekend) {
-        var r = getSatSun();
-        var sat = r[0];
-        var sun = r[1];
-        if (sat >= 0) {
-            t +=
-                tp +
-                '>div:nth-child(' +
-                (sat + 1) +
-                ') article:not(.registeredProgs) .w7_mH{background-color:rgba(227,238,255,0.7);}';
-            t +=
-                tp +
-                '>div:nth-child(' +
-                (sat + 1) +
-                ') article:not(.registeredProgs) .w7_mH:hover{background-color:rgba(222,233,250,0.7);}';
-        }
-        if (sun >= 0) {
-            t +=
-                tp +
-                '>div:nth-child(' +
-                (sun + 1) +
-                ') article:not(.registeredProgs) .w7_mH{background-color:rgba(255,227,238,0.7);}';
-            t +=
-                tp +
-                '>div:nth-child(' +
-                (sun + 1) +
-                ') article:not(.registeredProgs) .w7_mH:hover{background-color:rgba(250,222,233,0.7);}';
-        }
-    }
-
-    selBodyS = dl.getElementSingleSelector(EXTTbodyS);
-    if ($(selBodyS).length != 1) {
-        console.log('?EXTTbodyS ' + selBodyS);
-        selBody = alt ? '.i__jT' : '';
-    }
-    if (selBodyS) t += selBodyS + '{user-select:none;}'; //選択テキストを掴むと移動できないので選択不可にしておく
-
-    $(
-        "<link title='usermade' rel='stylesheet' href='data:text/css," +
-            encodeURIComponent(t) +
-            "'>"
-    ).appendTo('head');
-}
-function toggleChannel(targetUrl) {
-    //console.log("toggleChannel url="+targetUrl);
-    var t = /\/([^/]+)$/.exec(targetUrl)[1];
-    if (t == 'timetable') {
-        //ALLを選択した時
-        gdl.toast('番組表に表示するチャンネルをリセットしました。');
-        settings.allowChannelNames = [];
-        allowChannelNum = [];
-    } else {
-        var i = settings.allowChannelNames.indexOf(t);
-        var chname = getInfo.getChannelNameOnTimetable(t, EXTTsideL);
-        if (i < 0) {
-            //追加
-            gdl.toast(chname + 'を番組表に表示するチャンネルに追加しました。');
-            settings.allowChannelNames.push(t);
-        } else {
-            //削除
-            gdl.toast(
-                chname + 'を番組表に表示するチャンネルから削除しました。'
-            );
-            settings.allowChannelNames.splice(i, 1);
-        }
-    }
-    console.log(settings.allowChannelNames);
-    gl.setStorage({ allowChannelNames: settings.allowChannelNames.join(',') });
-    waitforloadtimetable(currentLocation);
-}
-function waitforloadtimetable(url) {
-    var c = getInfo.determineUrl();
-    if (c != getInfo.URL_DATETABLE && c != getInfo.URL_CHANNELTABLE) {
-        clearInterval(timetableRunning);
-        timetableRunning = false;
-        return;
-    }
-    if (url != currentLocation) return;
-    // 10分インターバル
-    if (timetableRunning === false) {
-        timetableRunning = setInterval(waitforloadtimetable, 600000, url);
-    }
-
-    var dd = $('div');
-    var alt = false;
-    $('head>link[title="usermade"]').remove(); //要素がdisplay:noneだと探索で大きさが取れないのでまず元に戻す
-    var j = dd.map(function(i, e) {
-        //ヘッダ探索、上の方にあって横長くて列の数は7日分くらいより多いやつ
-        var b = e.getBoundingClientRect();
-        if (
-            b.top < window.innerHeight / 4 &&
-            b.top > 5 &&
-            b.left < window.innerWidth / 3 &&
-            b.width > window.innerWidth / 2 &&
-            b.height < window.innerHeight / 3 &&
-            e.childElementCount > 5
-        )
-            return e;
-    });
-    if (j.length > 0) EXTThead = j[0];
-    else if (alt) {
-        j = $('.rT_r1').children('div');
-        if (j.length > 0) EXTThead = j[0];
-    }
-    if (!EXTThead) {
-        console.log('?EXTThead');
-        EXTThead = null;
-    }
-
-    j = dd.map(function(i, e) {
-        var b = e.getBoundingClientRect(); //ボディ探索、でかいやつ(ただしチャンネル選択時に幅を調整するので幅では判定しない)でdiv>div>articleがあるやつ left判定で左CHリスト幅の226を引く
-        if (
-            b.top < window.innerHeight / 4 &&
-            b.left - 226 < window.innerWidth / 4 &&
-            b.height > window.innerHeight / 2 &&
-            $(e).children('div').length > 5 &&
-            $(e).find('div>div>article').length > 0
-        )
-            return e;
-    });
-    if (j.length >= 0) EXTTbody = j[0];
-    else if (alt) {
-        j = $('.pi_pk');
-        if (j.length > 0) EXTTbody = j[0];
-    }
-    if (!EXTTbody) {
-        console.log('?EXTTbody');
-        EXTTbody = null;
-    }
-
-    j = dd.map(function(i, e) {
-        //番組詳細探索、右にあるやつ
-        var b = e.getBoundingClientRect();
-        if (
-            $(e).css('position') == 'fixed' &&
-            !isNaN(parseInt($(e).css('z-index'))) &&
-            parseInt($(e).css('z-index')) > 0 &&
-            b.top < window.innerHeight / 4 &&
-            b.left > window.innerWidth / 2 &&
-            b.width < window.innerWidth / 2 &&
-            b.width > 50 &&
-            b.height > (window.innerHeight * 2) / 3
-        )
-            return e;
-    });
-    if (j.length > 0) EXTTsideR = j[0];
-    else if (alt) {
-        j = $('.rT_sL');
-        if (j.length > 0) EXTTsideR = j[0];
-    }
-    if (!EXTTsideR) {
-        console.log('?EXTTsideR');
-        EXTTsideR = null;
-    }
-
-    j = $('a[href$="/timetable"]'); // 番組表リンク(ALL)をsideLとする
-    if (j.length >= 2) {
-        j = j.eq(1); //ヘッダの番組表リンクは除く
-        while (
-            j.prop('tagName').toUpperCase() != 'BODY' &&
-            j.children().length < 5
-        )
-            j = j.parent();
-        if (j.children().length >= 5) EXTTsideL = j[0];
-    }
-    if (!EXTTsideL && alt) {
-        j = $('.rT_sq ul');
-        if (j.length > 0) EXTTsideL = j[0];
-    }
-    if (!EXTTsideL) {
-        console.log('?EXTTsideL');
-        EXTTsideL = null;
-    }
-
-    if (EXTTbody) {
-        j = $(EXTTbody);
-        while (
-            j.prop('tagName').toUpperCase() != 'BODY' &&
-            j.css('overflow') == 'visible' &&
-            !j.siblings().is(EXTTsideL)
-        )
-            j = j.parent();
-        if (j.css('overflow') != 'visible') EXTTbodyS = j[0];
-    }
-    if (!EXTTbodyS && alt) {
-        j = $('.rT_ss');
-        if (j.length > 0) EXTTbodyS = j[0];
-    }
-    if (!EXTTbodyS) {
-        console.log('?EXTTbodyS');
-        EXTTbodyS = null;
-    }
-
-    j = dd.map(function(i, e) {
-        //timetable-axis
-        var b = e.getBoundingClientRect();
-        if (
-            b.top < window.innerHeight / 3 &&
-            b.left < window.innerWidth / 3 &&
-            b.width < 50 &&
-            b.height > (window.innerHeight * 2) / 3 &&
-            e.childElementCount > 20
-        )
-            return e;
-    });
-    if (j.length > 0) EXTTtime = j[0];
-    else if (alt) {
-        j = $('.pi_eR');
-        if (j.length > 0) EXTTtime = j[0];
-    }
-    if (!EXTTtime) {
-        console.log('?EXTTtime');
-        EXTTtime = null;
-    }
-
-    if (
-        EXTThead != null &&
-        EXTTbody != null &&
-        EXTTsideR != null &&
-        EXTTsideL != null &&
-        EXTTtime != null &&
-        EXTTbodyS != null
-    ) {
-        addExtClass(EXTThead, 'tt-head');
-        addExtClass(EXTTbody, 'tt-body');
-        addExtClass(EXTTsideR, 'tt-side-r');
-        addExtClass(EXTTsideL, 'tt-side-l');
-        addExtClass(EXTTtime, 'tt-time');
-        addExtClass(EXTTbodyS, 'tt-body-s');
-        allowChannelNumMaker();
-        if (c === getInfo.URL_CHANNELTABLE) {
-            setTimeout(timetablechfix, 100);
-        } else if (c === getInfo.URL_DATETABLE) {
-            setTimeout(timetabledtfix, 100);
-        }
-        setTimeout(timetableCommonFix, 100);
-        //番組表クリックで右詳細に通知登録ボタン設置
-        $(EXTTbody).click(function(e) {
-            // 掴んでスクロールした場合番組詳細は開かないことにする
-            if (!timetableGrabbing.scrolled) {
-                setTimeout(function() {
-                    var jSideR1 = $(EXTTsideR);
-                    var sideDetailDivClass = jSideR1.children().attr('class');
-                    var jSideR2 = jSideR1.siblings(
-                        ':has(div.' + sideDetailDivClass + ')'
-                    );
-                    if (jSideR1.css('z-index') < jSideR2.css('z-index')) {
-                        EXTTsideR = jSideR2[0];
-                        addExtClass(EXTTsideR, 'tt-side-r');
-                    }
-                    console.log(EXTTsideR);
-                    notifyButton.putSideDetailNotifyButton(
-                        settings.notifySeconds,
-                        EXTTsideL,
-                        EXTTsideR
-                    );
-                    if (settings.isPutSideDetailHighlight) {
-                        putSideDetailHighlight();
-                    }
-                    //右詳細が溢れてもスクロールできるように
-                    $(EXTTsideR).css('overflow-y', 'auto');
-                }, 100);
-
-                //console.log("putSideDetail*");
-            } else {
-                e.preventDefault();
-                e.stopImmediatePropagation();
-            }
-            //console.log('EXTTbody clicked',e,timetableGrabbing.scrolled)
-        });
-        timetableCss();
-        //$('div')を取ってきてあるのでここで使う 拡張のtoastとTimetableViewerスクリプトは除外する
-        dd.map(function(i, e) {
-            if (
-                $(e).css('z-index') > 10 &&
-                e.className.indexOf('ext-toast') < 0 &&
-                e.id.indexOf('TimetableViewer') < 0
-            )
-                return e;
-        }).css('z-index', '-1');
-
-        //番組表を掴んでドラッグする
-        timetableGrabbing.test = false;
-        timetableGrabbing.value = false;
-        $(EXTTbodyS).mouseleave();
-        if (timetableGrabbing.test == false) {
-            $(EXTTbodyS)
-                .mouseleave(function() {
-                    timetableGrabbing.test = true;
-                    timetableGrabbing.value = false;
-                    timetableGrabbing.scrolled = false;
-                })
-                .mouseup(function(e) {
-                    timetableGrabbing.value = false;
-                })
-                .mousedown(function(e) {
-                    if (e.buttons == 1) {
-                        // 左クリックだけの場合掴む
-                        timetableGrabbing.value = true;
-                        timetableGrabbing.cx = e.clientX;
-                        timetableGrabbing.cy = e.clientY;
-                        timetableGrabbing.sx = $(EXTTbodyS).scrollLeft();
-                        timetableGrabbing.sy = $(EXTTbodyS).scrollTop();
-                        timetableGrabbing.scrolled = false;
-                    } else {
-                        //timetableGrabbing.value=false;
-                        //timetableGrabbing.scrolled=false;
-                    }
-                })
-                .mousemove(function(e) {
-                    // 掴んで少し移動したらスクロール
-                    if (timetableGrabbing.value == false || e.buttons != 1) {
-                        //timetableGrabbing.scrolled=false;
-                    } else {
-                        if (
-                            timetableGrabbing.scrolled == false &&
-                            (Math.abs(timetableGrabbing.cx - e.clientX) > 10 ||
-                                Math.abs(timetableGrabbing.cy - e.clientY) > 10)
-                        )
-                            timetableGrabbing.scrolled = true;
-                        if (timetableGrabbing.scrolled) {
-                            $(EXTTbodyS)
-                                .scrollLeft(
-                                    timetableGrabbing.sx +
-                                        timetableGrabbing.cx -
-                                        e.clientX
-                                )
-                                .scrollTop(
-                                    timetableGrabbing.sy +
-                                        timetableGrabbing.cy -
-                                        e.clientY
-                                );
-                        }
-                    }
-                });
-        }
-    } else {
-        console.log('retry waitforloadtimetable');
-        setTimeout(waitforloadtimetable, 500, url);
-    }
-}
-function putSideDetailHighlight() {
-    var sideDetailWrapper = $(EXTTsideR);
-    if (
-        sideDetailWrapper.length == 0 ||
-        sideDetailWrapper.offset().left > window.innerWidth - 50
-    )
-        return;
-    sideDetailWrapper
-        .css('overflow-x', '')
-        .find('p[class="addedHighlight"]')
-        .remove();
-    var fp = sideDetailWrapper.find('p'); //番組詳細,タイトル,日時,見逃し云々?
-    if (fp.length < 2) return;
-    var progTitle = fp.eq(1).text();
-
-    var selPTitle = getTTProgramTitleClass();
-    if (!selPTitle) return;
-    var searchTarget = $(EXTTbody).find('.' + selPTitle);
-    var highlightString = '';
-    for (var i = 0, t, s; i < searchTarget.length; i++) {
-        t = searchTarget.eq(i).text();
-        if (t != progTitle) continue;
-        highlightString = searchTarget
-            .eq(i)
-            .parentsUntil('p')
-            .parent('p')
-            .next('p')
-            .text();
-        break;
-    }
-    if (!highlightString) return;
-    $(
-        '<p class="addedHighlight" style="line-height:19px;font-size:12px;margin-top:20px;">' +
-            highlightString +
-            '</p>'
-    ).appendTo(sideDetailWrapper.children().last());
-    sideDetailWrapper.css('overflow-x', 'hidden');
-}
-function timetabledtfix() {
-    if (EXTThead.childElementCount > EXTTbody.childElementCount) {
-        console.log('retry timetabledtfix()');
-        setTimeout(timetabledtfix, 500);
-        return;
-    }
-    //console.log("timetabledtfix");
-    //日付別番組表
-    //今はオプション1つのみだがチャンネル別のコピー
-    var ce = false;
-    var t = '';
-    if (settings.isChTimetablePlaybutton) {
-        ce = true;
-    }
-    if (ce) {
-        timetabledtloop();
-    }
-    console.log(allowChannelNum);
-    var channelLink = $(EXTThead)
-        .children('a')
-        .eq(0);
-    var chLinkWidth = 176;
-    var isTimetableScroll = false;
-    if (settings.timetableScroll != '') {
-        channelLink = $(EXTThead).children(
-            'a[href$="/timetable/channels/' + settings.timetableScroll + '"]'
-        );
-        if (channelLink.length > 0) {
-            isTimetableScroll = true;
-        } else {
-            channelLink = $(EXTThead)
-                .children('a')
-                .eq(0);
-            console.warn(
-                'timetable scroll error. (channelLink not found: チャンネル名が正しくないか仕様変更)'
-            );
-        }
-    }
-    var chLinkIndex = channelLink.index();
-    chLinkWidth = $(EXTTbody)
-        .children()
-        .eq(chLinkIndex)
-        .outerWidth(); //channelLink.width();
-    var visibleChLinkIndex = chLinkIndex;
-    var axisWidth = $(EXTTtime).width();
-    let timetableWidth;
-    if (allowChannelNum.length > 0) {
-        if (isTimetableScroll) {
-            visibleChLinkIndex = 0;
-            for (var i = 0; i < allowChannelNum.length; i++) {
-                if (allowChannelNum[i] < chLinkIndex) {
-                    visibleChLinkIndex++;
-                } else {
-                    break;
-                }
-            }
-            $(EXTTbody)
-                .parent()
-                .parent()
-                .parent()
-                .parent()
-                .scrollLeft(
-                    Math.min(
-                        chLinkWidth * visibleChLinkIndex,
-                        chLinkWidth * allowChannelNum.length -
-                            $(EXTTbodyS).width() +
-                            axisWidth
-                    )
-                );
-        }
-        if (!settings.isExpandFewChannels) {
-            timetableWidth = axisWidth + chLinkWidth * allowChannelNum.length;
-        } else {
-            timetableWidth =
-                axisWidth + chLinkWidth * EXTThead.childElementCount;
-        }
-    } else {
-        if (isTimetableScroll) {
-            $(EXTTbody)
-                .parent()
-                .parent()
-                .parent()
-                .parent()
-                .scrollLeft(chLinkWidth * visibleChLinkIndex);
-        }
-        timetableWidth = axisWidth + chLinkWidth * EXTThead.childElementCount;
-    }
-    //番組表幅の調整
-    //console.log(timetableWidth)
-    $(EXTThead)
-        .parent()
-        .innerWidth(timetableWidth);
-    $(EXTTbodyS)
-        .children()
-        .width(timetableWidth)
-        .children('div')
-        .last()
-        .width(timetableWidth);
-
-    //左チャンネル一覧にチェックボックス設置
-    var channelsli = $(EXTTsideL).children('li');
-    if (channelsli.length == 0) {
-        console.warn('channelsli');
-    }
-    channelsli.each(function(i, li) {
-        // if(i == 0){return;}
-        // i--;
-        li = $(li);
-        var checkbox = li.children('input');
-        if (checkbox.length == 0) {
-            checkbox = $(
-                '<input type="checkbox" class="usermade chlicheckbox" style="display:inline-block;margin:' +
-                    (gl.isFirefox ? 7 : 8) +
-                    'px;height:12px;vertical-align:middle;" title="拡張機能のチャンネル表示切替">'
-            ).appendTo(li);
-            checkbox.click(function(e) {
-                toggleChannel(
-                    e.currentTarget.previousElementSibling.getAttribute('href')
-                );
-            });
-        }
-        if (i == 0) {
-            checkbox.prop('checked', allowChannelNum.length == 0);
-            checkbox.prop('disabled', allowChannelNum.length == 0);
-        } else {
-            checkbox.prop('checked', gl.hasArray(allowChannelNum, i - 1));
-        }
-        li.children('a')
-            .css('display', 'inline-block')
-            .css('width', 'calc(100% - ' + (16 + checkbox.width()) + 'px)');
-    });
-}
-function timetabledtloop() {
-    if (getInfo.determineUrl() != getInfo.URL_DATETABLE) return;
-    if (!settings.isChTimetablePlaybutton) return;
-    if (settings.isChTimetablePlaybutton) {
-        PlaybuttonEditor();
-    }
-    setTimeout(timetabledtloop, 1000);
-}
-function timetablechfix() {
-    //console.log("timetablechfix");
-    //チャンネル別番組表
-    var ce = false; //定期実行するかどうか
-
-    if (settings.isChTimetablePlaybutton) {
-        //再生ボタンの設置場所(放送中の緑色枠)は移動するので定期実行にて削除,設置する
-        ce = true;
-    }
-
-    if (ce) {
-        timetablechloop();
-    }
-}
-function getSatSun() {
-    var h = $(EXTThead).children('a[href*="/timetable/dates/"]');
-    var sat = -1;
-    var sun = -1;
-    for (var i = 0; i < h.length; i++) {
-        if (/[(（]土[)）]/.test(h.eq(i).text())) {
-            sat = i;
-            if (i < h.length - 1) {
-                sun = i + 1;
-                break;
-            }
-        } else if (/[(（]日[)）]/.test(h.eq(i).text())) {
-            sun = i;
-            if (i > 0) {
-                sat = i - 1;
-                break;
-            }
-        } else if (/[(（]月[)）]/.test(h.eq(i).text())) {
-            if (i > 0) {
-                sun = i - 1;
-                if (i - 1 == 0) {
-                    break;
-                } else if (i > 1) {
-                    sat = i - 2;
-                    break;
-                }
-            }
-        }
-    }
-    return [sat, sun];
-}
-function timetablechloop() {
-    //URL変わったら終われるようにURLチェック
-    if (getInfo.determineUrl() != getInfo.URL_CHANNELTABLE) return;
-    if (!settings.isChTimetablePlaybutton) return;
-    if (settings.isChTimetablePlaybutton) {
-        PlaybuttonEditor();
-    }
-    setTimeout(timetablechloop, 1000);
-}
-function PlaybuttonEditor() {
-    if (!settings.isChTimetablePlaybutton) return;
-    if (/^https:\/\/abema\.tv\/timetable\/dates\/.+$/.test(currentLocation))
-        return; //当日を除く日付別番組表では実行しない
-    //放送中の緑枠のclassを取得
-    var fisrtChDivs = EXTTbody.firstElementChild.childNodes;
-    var presentClass = '';
-    var clsArr = [];
-    //一つのチャンネルの一日分番組divを取得しそのarticle>button>divのclass名が仲間はずれのものが放送中のclass
-    for (var i = 0, cls, flg = true; i < fisrtChDivs.length; i++) {
-        cls = $(fisrtChDivs[i])
-            .find('article>button>div')
-            .attr('class');
-        flg = true;
-        for (var j = 0; j < clsArr.length; j++) {
-            if (clsArr[j][0] == cls) {
-                clsArr[j][1]++;
-                flg = false;
-            }
-        }
-        if (flg) {
-            clsArr.push([cls, 1]);
-        }
-    }
-    if (clsArr.length == 3 && clsArr[1][1] == 1) {
-        presentClass = clsArr[1][0];
-    } else if (clsArr.length == 2) {
-        if (clsArr[0][1] == 1) presentClass = clsArr[0][0];
-        else if (clsArr[1][1] == 1) presentClass = clsArr[1][0];
-    }
-    if (presentClass.length == 0) {
-        console.warn('?presentClass failed');
-        return;
-    }
-    var presentSelector = '.' + presentClass.split(' ').join('.');
-    //放送中の緑枠の移動に合わせて再生ボタンを削除、設置する
-    var p = $(presentSelector); //放送中の緑色枠
-    var b = $('.playbutton');
-    var c = $(EXTThead).children(); //channel link
-    var cr = /^https:\/\/abema\.tv\/timetable\/channels\/(.+)$/;
-    var dr = /^https:\/\/abema\.tv\/timetable(?:\/dates\/.+)?$/;
-    var umc = currentLocation.match(cr);
-    var umd = currentLocation.match(dr);
-    var pn = -1;
-    var bn = -2;
-    var titleClass = getTTProgramTitleClass();
-    for (let i = b.length - 1, d, s; i >= 0; i--) {
-        d = b.eq(i).parent('a');
-        s = d.siblings();
-        if (!s.is(presentSelector)) {
-            //設置済ボタン位置が緑枠でなければボタン削除
-            s.find('.' + titleClass)
-                .parents('p')
-                .css('width', '');
-            d.remove();
-        }
-    }
-    for (let i = 0, j, q, a, u, iumc; i < p.length; i++) {
-        q = p.eq(i);
-        if (umc && umc.length > 1 && umc[1].length > 0) {
-            //チャンネル別番組表ならボタンのリンク先はURLから取得
-            u = '/now-on-air/' + umc[1];
-        } else if (umd && umd.length > 0 && c.length > 0) {
-            //日付別番組表ならアイコンのリンクから取得
-            j = $(EXTTbody)
-                .children()
-                .has(q)
-                .index(); //列のindex
-            iumc = c
-                .eq(j)
-                .prop('href')
-                .match(cr);
-            if (iumc && iumc.length > 1 && iumc[1].length > 0) {
-                u = '/now-on-air/' + iumc[1];
-            } else {
-                u = '';
-            }
-        } else {
-            //何か変な場合はトップページへ飛ぶようにする
-            u = '';
-        }
-        if (
-            !q
-                .siblings()
-                .children()
-                .is('.playbutton')
-        ) {
-            //緑枠にボタンがなければ設置
-            q.find('.' + titleClass)
-                .parents('p')
-                .css('width', '86px');
-            a =
-                '<a href="javascript:location.href=\'https://abema.tv' +
-                u +
-                '\';" title="放送中画面へ移動">';
-            a +=
-                '<div class="playbutton" style="position:absolute;right:30px;top:4px;width:24px;height:24px;border:1px solid #6fb900;border-radius:50%;">';
-            a +=
-                '<svg width="10" height="14" style="fill:#6fb900;transform:translate(1px,3px)">'; // 以前は7px,3px
-            a += '<use xlink:href="/images/icons/playback.svg#svg-body">';
-            a += '</use></svg></div>';
-            a += '</a>';
-            $(a).insertAfter(q);
-            $('.playbutton').on('click', function(e) {
-                //普通の左クリックのみ移動、特殊クリックの場合はその操作に従う(移動しない)
-                //console.log(e);
-                if (
-                    e.which == 1 &&
-                    e.altKey == false &&
-                    e.ctrlKey == false &&
-                    e.shiftKey == false
-                ) {
-                    e.preventDefault();
-                    e.stopImmediatePropagation();
-                    //再生ボタンのある番組をクリックして右詳細の番組画像をクリック
-                    //console.log(e.currentTarget);
-                    clickElement($(e.currentTarget).parents('button'));
-                    setTimeout(
-                        clickElement,
-                        10,
-                        $(EXTTsideR).find('a[href^="/now-on-air/"]')
-                    ); //todo
-                }
-            });
-        }
-    }
-}
-
-function clickElement(jo) {
-    //clickOnairLink(jo)
-    //console.log("clickElement", jo);
-    var teka = document.createEvent('MouseEvents');
-    teka.initMouseEvent('click', true, true, window);
-    //console.log("dispatch");
-    return jo[0].dispatchEvent(teka);
-}
-
-function timetableCommonFix(retrycount) {
-    var progArticle, progTitle;
-    //読込に1秒以上かかる場合を考慮し、ヘッダと列の数を見るようにしてみる
-    var cols = $(EXTTbody)
-        .children('div')
-        .map(function(i, e) {
-            if (
-                $(e).width() > 50 &&
-                $(e).height() > (window.innerHeight * 2) / 3
-            )
-                return e;
-        });
-
-    if (cols.length < EXTThead.childElementCount) {
-        setTimeout(timetableCommonFix, 1000);
-        return;
-    }
-
-    //別の日のページへの遷移直後は前の読込済表示が残っている(cols.len=head.childcount)ので更新を待つ必要がある
-    //とりあえず最後の番組のarticleにtitleが付いてたら残ってることにしてリトライする
-    if (
-        cols
-            .last()
-            .children()
-            .last()
-            .find('article')
-            .attr('title')
-    ) {
-        if (retrycount === undefined) retrycount = 5;
-        if (retrycount > 0)
-            setTimeout(timetableCommonFix, 1000, retrycount - 1);
-    }
-
-    //番組タイトルをtitle要素にする
-    var selPTitle = getTTProgramTitleClass();
-    var selICont = getTTProgramTimeClasses()[1];
-    var jt, jp, jf;
-    cols.each(function() {
-        $(this)
-            .children()
-            .each(function() {
-                //番組毎divについてのループ
-                progArticle = $(this).find('article');
-                jt = progArticle.find('.' + selPTitle);
-                progTitle = jt.text();
-                progArticle.attr('title', progTitle);
-
-                if (settings.isReplaceIcons && selICont) {
-                    jp = jt.siblings('span,svg');
-                    if (jp.length > 0) {
-                        jf = progArticle.find('.' + selICont);
-                        if (jf.length > 0) {
-                            jp.each(function() {
-                                $(this).appendTo(jf);
-                            });
-                        }
-                    }
-                }
-            });
-    });
-    notifyButton.setRegistProgsBackground();
-}
-
 function onresize(oldTranslate) {
     if (getInfo.determineUrl() != getInfo.URL_ONAIR) return;
     //console.log("onresize()");
@@ -1641,7 +696,16 @@ function putComment(commentText, userid, index, inmax, isSelf) {
         kakikomitxt = '';
         // console.log("kakikomitxt reset: putComment")
     }
-    commentText = gcl.comefilter(commentText, userid, arFullNg, arUserNg, settings.isComeDel, settings.isUserDel, settings.isComeNg, settings.isDeleteStrangeCaps);
+    commentText = gcl.comefilter(
+        commentText,
+        userid,
+        arFullNg,
+        arUserNg,
+        settings.isComeDel,
+        settings.isUserDel,
+        settings.isComeNg,
+        settings.isDeleteStrangeCaps
+    );
     var commentTopMargin = 50;
     var commentBottomMargin = 150;
     if (EXhead.style.visibility == 'hidden') {
@@ -1963,7 +1027,6 @@ function openOption() {
     $('#timetableScroll').val(settings.timetableScroll);
     $('#isHideTwitterPanel').prop('checked', settings.isHideTwitterPanel);
     $('#isHideTodayHighlitht').prop('checked', settings.isHideTodayHighlight);
-    //$('#isComelistNG').prop("checked", settings.isComelistNG);
     $('#isComelistClickNG').prop('checked', settings.isComelistClickNG);
     $(
         '#ihighlightComeColor input[type="radio"][name="highlightComeColor"]'
@@ -2062,8 +1125,7 @@ function closeOption() {
         .css('right', '40px');
     $('.rightshift').css('display', 'none');
     $('.leftshift').css('display', '');
-    var jo = $(EXcomelist)
-        .children('div');
+    var jo = $(EXcomelist).children('div');
     jo.css('background-color', '') //基本色、新着強調
         .css('color', '') //基本色
         .css('border-left', '') //新着強調
@@ -2205,7 +1267,7 @@ function delayset(
 
     var resetOptionHead = false;
     if (!isInfo && (EXinfo = getInfoElement())) {
-        addExtClass(EXinfo, 'info');
+        dl.addExtClass(EXinfo, 'info');
         console.log('setOptionHead delayset(EXinfo)');
         resetOptionHead = true;
         isInfo = true;
@@ -2225,7 +1287,7 @@ function delayset(
         }
     }
     if (!isChli && (EXchli = getElm.getChannelListElement())) {
-        addExtClass(EXchli, 'channelList');
+        dl.addExtClass(EXchli, 'channelList');
         console.log('setOptionHead delayset(EXchli)');
         resetOptionHead = true;
         isChli = true;
@@ -2243,7 +1305,7 @@ function delayset(
             );
     }
     if (!isComeli && (EXcomelist = getComeListElement())) {
-        addExtClass(EXcomelist, 'comelist');
+        dl.addExtClass(EXcomelist, 'comelist');
         window.dispatchEvent(comelistReadyEvent);
         console.log('setOptionHead delayset(EXchli)');
         resetOptionHead = true;
@@ -2255,7 +1317,7 @@ function delayset(
         isTwT = true;
     }
     if (!isEXC && $(EXcomelist).length > 0) {
-        //        addExtClass(EXcomelist, 'comelist');
+        //        dl.addExtClass(EXcomelist, 'comelist');
         setTimeout(applyCommentListNG, 1000);
 
         //コメ欄のDOM変更監視
@@ -2349,13 +1411,11 @@ function delaysetNotOA() {
         .className;
     //console.log(hoverContents,hoverContents.children(),hoverLinkClass)
     if (hoverContents.children('#extSettingLink').length == 0) {
-        hoverContents
-            .children(':last')
-            .css({
-                'border-bottom': '1px solid #333',
-                'margin-bottom': '8px',
-                'padding-bottom': '12px'
-            });
+        hoverContents.children(':last').css({
+            'border-bottom': '1px solid #333',
+            'margin-bottom': '8px',
+            'padding-bottom': '12px'
+        });
         hoverContents
             .append(
                 '<a class="' +
@@ -3690,7 +2750,6 @@ function setSaveClicked() {
     //    settings.isChTimetablePlaybutton=$('#isChTimetablePlaybutton').prop("checked");
     settings.isHideTwitterPanel = $('#isHideTwitterPanel').prop('checked');
     settings.isHideTodayHighlight = $('#isHideTodayHighlight').prop('checked');
-    //settings.isComelistNG = $('#isComelistNG').prop("checked");
     settings.isComelistClickNG = $('#isComelistClickNG').prop('checked');
     settings.highlightComeColor = parseInt(
         $(
@@ -4484,12 +3543,7 @@ function comemarginfix(repeatcount, inptime, inptitle, inpsame, inpbig) {
         );
     }
 }
-function addExtClass(elm, className) {
-    //return ;
-    className = 'ext_abm-' + className;
-    $('.' + className).removeClass(className);
-    $(elm).addClass(className);
-}
+
 function setEXs() {
     //ロード直後に取得が期待できるやつ
     //obliが遅くinfoはもっと遅い
@@ -4570,23 +3624,23 @@ function setEXs() {
 
     if (b == true) {
         // class付与
-        addExtClass(EXhead, 'header');
-        addExtClass(EXmenu, 'menu');
-        addExtClass(EXfoot, 'footer');
-        addExtClass(EXfootcome, 'footcome');
-        addExtClass(EXcountview, 'countview');
-        addExtClass(EXfootcountcome, 'footcountcome');
-        addExtClass(EXside, 'sideButton');
-        //addExtClass(EXchli, "channelList");
-        //addExtClass(EXinfo, 'info');
-        addExtClass(EXcome, 'come');
-        addExtClass(EXcomesend, 'comesend');
-        addExtClass(EXcomesendinp, 'comesendinput');
-        addExtClass(EXvolume, 'volume');
-        addExtClass(EXfullscr, 'fullscr');
-        //addExtClass(EXobli, 'objectlist');
-        addExtClass(EXvideoarea, 'videoarea');
-        addExtClass(EXcomelist, 'comelist');
+        dl.addExtClass(EXhead, 'header');
+        dl.addExtClass(EXmenu, 'menu');
+        dl.addExtClass(EXfoot, 'footer');
+        dl.addExtClass(EXfootcome, 'footcome');
+        dl.addExtClass(EXcountview, 'countview');
+        dl.addExtClass(EXfootcountcome, 'footcountcome');
+        dl.addExtClass(EXside, 'sideButton');
+        //dl.addExtClass(EXchli, "channelList");
+        //dl.addExtClass(EXinfo, 'info');
+        dl.addExtClass(EXcome, 'come');
+        dl.addExtClass(EXcomesend, 'comesend');
+        dl.addExtClass(EXcomesendinp, 'comesendinput');
+        dl.addExtClass(EXvolume, 'volume');
+        dl.addExtClass(EXfullscr, 'fullscr');
+        //dl.addExtClass(EXobli, 'objectlist');
+        dl.addExtClass(EXvideoarea, 'videoarea');
+        dl.addExtClass(EXcomelist, 'comelist');
 
         console.log('%csetEXs ok', 'color:green;');
         //setEX2(); 残ってたchli.scrollをdelaysetに移動させてsetex2を空にした
@@ -5107,248 +4161,7 @@ function getVideoRouteClasses() {
             .getAttribute('class')
     ];
 }
-function getTTProgramTitleClass() {
-    // 0:ビデオのN、1～その他、10くらいから番組タイトル
-    //後ろから取るが、番組表の後ろに要素が新設されても対応できるように適当なオフセットを設けておく
-    return $(EXTTbody)
-        .find('span')
-        .map(function(i, e) {
-            if (
-                e.childElementCount == 0 &&
-                e.className != '' &&
-                e.textContent != ''
-            )
-                return e;
-        })
-        .eq(-5)
-        .prop('class');
-}
-function getTTTimeClassesFromPT(proTitleClass) {
-    // return[] 0:過去 1:放送中 2:放送予定 の背景を司るクラスを番組タイトルクラスから探す
-    // programtitleの親の(articleの子の)buttonの子のdivが灰/緑/白の背景を持っているのでそれを探す
-    // '.'を付けたクラス名で受ける(ここでは付けない)
-    var ret = [null, null, null];
-    if (!proTitleClass) {
-        proTitleClass = getTTProgramTitleClass();
-        if (!proTitleClass) return ret;
-        else proTitleClass = '.' + proTitleClass;
-    }
 
-    var jb = $(EXTTbody);
-    var jo = jb.find(proTitleClass);
-    var re = /rgba?\( *(\d+) *, *(\d+) *, *(\d+)(?: *, *\d+ *,?)? *\)/;
-    var rs = /\s/;
-    var rr = /^\s+|\s+$/g;
-    var classes = [null, null, null];
-    for (var i = 0, j, t, e, r, g, b, c; i < jo.length; i++) {
-        j = jo
-            .eq(i)
-            .parentsUntil('button')
-            .last();
-        t = j.css('background-color');
-        if (!re.test(t)) continue;
-        e = re.exec(t);
-        r = parseInt(e[1]);
-        g = parseInt(e[2]);
-        b = parseInt(e[3]);
-        if (!classes[0] && r < 248 && g < 248 && b < 248) c = 0;
-        else if (!classes[1] && r < 248 && g > 248 && b < 248) c = 1;
-        else if (!classes[2] && r > 248 && g > 248 && b > 248) c = 2;
-        else continue;
-        classes[c] = j.prop('class').split(rs);
-        if (classes[0] && classes[1] && classes[2]) break;
-    }
-
-    // timetable/dates/など全部過去、全部未来の場合(1つしか取れてない)はクラスが該当する要素が同数で判別できない
-    var nc = 0;
-    for (let i = 0; i < 3; i++) if (!classes[i]) nc++;
-    if (nc == 3 || nc == 2) return ret;
-    //(2以上の要素内で)重複クラスを削除して1つにならなければ全体の該当要素が少ないのを選ぶ
-    for (let i = 0; i < 3; i++) {}
-    var jc = 9999;
-    var eq = true;
-    ret = null;
-    for (let i = 0, ci, cl; i < classes.length; i++) {
-        ci = classes[i].replace(rr, '');
-        if (!ci) continue;
-        cl = jb.find('.' + ci).length;
-        if (jc > cl) {
-            jc = cl;
-            ret = ci;
-            eq = i == 0;
-        } else if (jc < cl) eq = false;
-    }
-    return eq ? null : ret;
-}
-function getTTLRArrowContainerElement(returnSingleSelector) {
-    //右にある右アイコンの親buttonの親divを選びたいが初期状態では存在せず取れない場合があるので、
-    //横に長くて縦が短く画面中央にあってtimebarでないものを選ぶ
-    //var jo = $('[*|href$="/images/icons/chevron_right.svg#svg-body"]:not([href])');
-    var jo = $('div').map(function(i, e) {
-        if (
-            e.clientWidth > window.innerWidth / 2 &&
-            e.offsetTop > window.innerHeight / 3 &&
-            e.offsetTop < (window.innerHeight * 2) / 3
-        )
-            return e;
-    });
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    //特定条件でtimebarも取れるので除外する
-    var ret = getTTtimebarElement();
-    if (ret) {
-        jo = jo.not(ret);
-        if (jo.length == 0) return null;
-        else if (jo.length == 1)
-            return returnSingleSelector
-                ? dl.getElementSingleSelector(jo[0])
-                : jo[0];
-    }
-    for (let i = 0; i < jo.length; i++) {
-        if ($(EXTTtime).find(jo.eq(i)).length > 0) jo = jo.not(jo.eq(i));
-    }
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    for (let i = 0, z; i < jo.length; i++) {
-        z = jo.eq(i).css('z-index');
-        if (z == 'auto' || isNaN(parseInt(z)) || parseInt(z) <= 0)
-            jo = jo.not(jo.eq(i));
-    }
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    for (let i = 0, z; i < jo.length; i++) {
-        if (jo.eq(i).offset().left > 50) jo = jo.not(jo.eq(i));
-    }
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    return null;
-}
-function getTTtimebarElement(returnSingleSelector) {
-    //横に長くて縦が短くtopが直接指定されてるのを選ぶ
-    let jo = $('div').map(function(i, e) {
-        if (
-            e.clientWidth > window.innerWidth / 2 &&
-            e.clientHeight < 30 &&
-            e.style.top != ''
-        )
-            return e;
-    });
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    //もし2以上引っかかったら時刻を探す
-    var re = /(^|[^\d])\d{1,2}:\d{2}($|[^\d])/;
-    jo = jo.map(function(i, e) {
-        if (re.test(e.textContent)) return e;
-    });
-    if (jo.length == 0) return null;
-    else if (jo.length == 1)
-        return returnSingleSelector
-            ? dl.getElementSingleSelector(jo[0])
-            : jo[0];
-
-    //この時点でまだ取りきれないなら背景を使う
-    var rt = /rgba? *\( *(\d+) *, *(\d+) *, *(\d+)(?: *,\d+ *,?)? *\)/;
-    jo = $('p').map(function(i, e) {
-        if (e.offsetHeight > 0 && e.offsetHeight < 30 && re.test(e.textContent))
-            return e;
-    });
-    var ret = null;
-    for (var i = 0, c, e, r, g, b; i < jo.length; i++) {
-        c = jo.eq(i).css('background-color');
-        if (!rt.test(c)) continue;
-        e = rt.exec(c);
-        r = parseInt(e[1]);
-        g = parseInt(e[2]);
-        b = parseInt(e[3]);
-        if ((r < 192 && g < 192 && b < 192) || (r > 64 && g > 64 && b > 64))
-            continue;
-        ret = jo[i];
-        break;
-    }
-    if (!ret) return null;
-    var p = ret.parentElement;
-    while (p.tagName.toUpperCase() != 'BODY' && p.offsetHeight < 30) {
-        ret = p;
-        p = ret.parentElement;
-    }
-    if (p.tagName.toUpperCase() == 'BODY') return null;
-    return returnSingleSelector ? dl.getElementSingleSelector(ret) : ret;
-}
-function getTTProgramTimeClasses() {
-    //番組開始の00とか30のクラス とその中のアイコンコンテナ(FREEとかを収納する用)があれば取る
-    //数字2桁をexttbody以外から取ると時間軸と日付も引っかかる
-    var ret = [null, null];
-    var jo = $(EXTTbody)
-        .find('div')
-        .map(function(i, e) {
-            if (/^\d{2}$/.test(e.textContent)) return e;
-        });
-    var ja = [];
-    for (var i = 0, c, added; i < jo.length; i++) {
-        c = jo.eq(i).prop('class');
-        if (!/\w/.test(c)) continue;
-        c = c.split(/\s/)[0].replace(/^\s+|\s+$/, '');
-        added = false;
-        for (var j = 0; j < ja.length; j++) {
-            if (ja[j][0] != c) continue;
-            ja[j][1]++;
-            added = true;
-            break;
-        }
-        if (!added) {
-            ja[ja.length] = [c, 1];
-        }
-    }
-    if (ja.length == 0) return ret;
-    var t = ja[0][0];
-    var m = ja[0][1];
-    for (let i = 1; i < ja.length; i++) {
-        if (m > ja[i][1]) continue;
-        t = ja[i][0];
-        m = ja[i][1];
-    }
-    ret[0] = t;
-
-    //時刻の後ろでtable-cellなdivのclassを選ぶ
-    jo = $('.' + t);
-    for (let i = 0, ji; i < jo.length; i++) {
-        ji = jo.eq(i).contents();
-        for (let j = 0, jp; j < ji.length; j++) {
-            if (!/^\d{2}$/.test(ji.eq(j).text())) continue;
-            jp = ji.eq(j).nextAll('div');
-            for (let k = 0, c; k < jp.length; k++) {
-                if (jp.eq(k).css('display') != 'table-cell') continue;
-                c = jp.eq(k).prop('class');
-                if (!/\w/.test(c)) continue;
-                ret[1] = c.split(/\s/)[0].replace(/\s+|\s+$/, '');
-                break;
-            }
-            if (ret[1]) break;
-        }
-        if (ret[1]) break;
-    }
-    return ret;
-}
 function hasNotTransformed(jo) {
     //transformしていない=right0等が画面外に移動していない=開いている
     //1つでも開いてればtrue,全部閉じてたらfalse
@@ -7623,9 +6436,9 @@ function usereventFCclick() {
             }, 500); // コメ欄を開くと公式が映像サイズを縮めてしまうので広げ直す
         }
         // コメ欄が開かれるので全画面・音量ボタンずらす(黒帯トリミングの場合は不要)
-        if(!(settings.isComeTriming && settings.isSureReadComment)){
+        if (!(settings.isComeTriming && settings.isSureReadComment)) {
             EXfullscr.style.right = commentListWidth + fullscrRight + 'px';
-            EXvolume.style.right = commentListWidth + volumeRight + 'px';    
+            EXvolume.style.right = commentListWidth + volumeRight + 'px';
         }
     }
     var jo = $(getElm.getVideo());
@@ -8283,7 +7096,16 @@ function applyCommentListNG(d) {
             continue;
         commentDivs[i].setAttribute('data-ext-filtered', 'true');
         let cinfo = getComeInfo(commentDivs[i]);
-        let filteredComment = gcl.comefilter(cinfo.message, cinfo.userid, arFullNg, arUserNg, settings.isComeDel, settings.isUserDel, settings.isComeNg, settings.isDeleteStrangeCaps);
+        let filteredComment = gcl.comefilter(
+            cinfo.message,
+            cinfo.userid,
+            arFullNg,
+            arUserNg,
+            settings.isComeDel,
+            settings.isUserDel,
+            settings.isComeNg,
+            settings.isDeleteStrangeCaps
+        );
         //commentDivs[i].classList.add('comew');
         if (
             !(
@@ -8829,16 +7651,16 @@ function onairBasefunc() {
 
         // 1秒ごとに実行
         if ($('.ext_abm-come').length == 0) {
-            addExtClass(EXcome, 'come');
+            dl.addExtClass(EXcome, 'come');
         }
         if ($('.ext_abm-comelist').length == 0) {
-            addExtClass(EXcomelist, 'comelist');
+            dl.addExtClass(EXcomelist, 'comelist');
         }
         if (!EXcomelist || $(document).has(EXcomelist).length == 0) {
             EXcomelist = getComeListElement();
             if (EXcomelist) {
                 //console.log('ecl', EXcomelist, $('body').has(EXcomelist).length==0)
-                addExtClass(EXcomelist, 'comelist');
+                dl.addExtClass(EXcomelist, 'comelist');
                 setOptionHead();
                 window.dispatchEvent(comelistReadyEvent);
                 commentObserver.disconnect();
@@ -8847,7 +7669,7 @@ function onairBasefunc() {
         }
         if (!EXinfo || $(document).has(EXinfo).length == 0) {
             EXinfo = getInfoElement();
-            addExtClass(EXinfo, 'info');
+            dl.addExtClass(EXinfo, 'info');
             setOptionHead();
         }
 
@@ -9737,13 +8559,13 @@ function URLPatternSwitch() {
             //番組表に再生ボタンを追加する機能があるため、ここにあった放送画面へのリンクは廃止
             onairCleaner();
             delaysetNotOA();
-            waitforloadtimetable(url);
+            TT.waitforloadtimetable(url);
             break;
         case getInfo.URL_CHANNELTABLE:
             //チャンネル別番組表
             onairCleaner();
             delaysetNotOA();
-            waitforloadtimetable(url);
+            TT.waitforloadtimetable(url);
             break;
         case getInfo.URL_ONAIR:
             //放送ページ
@@ -9840,7 +8662,7 @@ chrome.runtime.onMessage.addListener(function(r) {
             addPermanentNG(r.word);
         }
     } else if (r.name == 'toggleChannel') {
-        toggleChannel(r.url);
+        TT.toggleChannel(r.url);
     } else if (r.name == 'historyStateUpdated') {
         chkurl();
     } else {
